requirements.txt
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
fastapi==0.115.8
filelock==3.17.0
fsspec==2025.2.0
h11==0.14.0
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
minio==7.2.15
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
pika==1.3.2
pillow==11.1.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
python-dotenv==1.0.1
python-multipart==0.0.20
setuptools==75.8.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
